Move constant.ts directory
